move NotificationTypeConverter into Converters
